src/main/webapp/app/entities/repayment/delete/repayment-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { RepaymentService } from '../service/repayment.service';

import { RepaymentDeleteDialogComponent } from './repayment-delete-dialog.component';

describe('Repayment Management Delete Component', () => {
  let comp: RepaymentDeleteDialogComponent;
  let fixture: ComponentFixture<RepaymentDeleteDialogComponent>;
  let service: RepaymentService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RepaymentDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(RepaymentDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(RepaymentDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(RepaymentService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
